[IMP] l10n_mx_edi_cfdi_test: Fixed test expected
The serie is added with the current year when the invoice was validated.

Then take the 2017, but now is 2018, by this reason was copied the serie
from the expected to the cfdi to compare, to always have the same
values.

Also, was moved the sign test to l10n_mx_edi_cfdi_33, because that
module have the correct system parameter to CFDI 3.3