[REF] web_enterprise: FormView: style for XXS screens
This rev. stops using media queries to define the style of form
views on XXS screens (width < 475px). Using media queries is a
limitation as the style of the view is determined by the size of
the screen, whether it is rather linked to the size of its
container.

For instance, a form view is rendered inside the QuickCreate widget
in kanban views. This form view has a width < 475px, even on large
screens, and we want the XXS style to apply on this view.

So instead of media queries, we now toggle classname
'o_xxs_form_view' on form views that require the XXS style (this
classname is automatically toggled on XXS screens, and hardcoded
in form views inside QuickCreate widgets).

Much of the form XXS rules have been moved to community, as they
are now also required in the the Kanban QuickCreate.