[IMP] l10n_mx_edi: Regenerate CFDI in all cases
When a invoice is validated, and the CFDI structure is correct, the
system send to the PAC the document to try stamp.

But if a value is incorrect:

Ex. The VAT do not exist

The PAC return the error, then if is fixed the VAT, and press Retry
button, the CFDI is not regenerated, then the invoice is must be
cancelled, reset2draft and resign to take the changes.

Now the case does not matter when you press the Retry button the CFDI
always must be regenerated.