[FIX] web_mobile: reactivate status buttons on mobile apps redirects
Before this commit, status buttons did not work anymore after
url redirect performed by the Odoo apps (Android and iOS).

For example, if you create a new quotation and you click
on the 'preview' buttons, you are redirected to the browser.
But if you want to go back the app, it is not longer possible
to click on the 'save' buttons (neither others).

In odoo, the form buttons are disabled after clicking on action
to avoid multiple writes. These buttons will be enable again
when the action is finished.
In this case, the deferred isn't resolved but the new url is
supposed to be opened in the same web page.
This is quite different with the mobile apps which force website
urls to be opened in mobile brower.

To fix this, we return a resolved deferred as if we open it in a new
window. The status buttons will be then reactivated.

In this case, the 'is_mobile' flag is not good enough to detect if
we are in a mobile app because this variable is also set to 'true'
for mobile browsers.

opw-775338