[FIX] web_mobile: many2one color in iOS
Before this commit, many2one fields were disabled to avoid to show
the autocomplete. In mobile app, the user has to click on the
input to see propositions.

But iOS has a specific css rule that changes the color of disabled
text html inputs. This was close to the style of a placeholder.
So the user didn't know if the value was set or not.

Disable fields were not the way to go and we can now change this
with the 'new views refactoring' because '_bindAutoComplete' is
now a separated method that we can easily override for mobile apps.

So, it's not necessary to disable manyone anymore and it fixes the
css Safary (iOS) color rule in the same time.