[FIX] account_reports{_followup}: lang of letter
- Install a second language (e.g. French)
- Set the French language to a partner A
- Create an overdue invoice for A
- From the partner form view, click on the "Due" stat button

The complete view is translated to French, including the buttons ('Print
Letter'...). There are also inconsistencies in the language of the
letter printed (PDF or email).

Loading the action with the partner lang is incorrect. We should instead
translate the elements of the followup letter at the appropriate moment
manually.

opw-788301