[FIX] web_enterprise: properly set MobileFormRenderer
In web_enterprise, we override the FormView to use a specific
Renderer in mobile. However, the way it was done didn't allow to
define extensions of the FormView (using specific Renderers).

We now reject the file doing the include on FormView if we are not
in mobile, and we thus moved two mobile form tests to the mobile
test suite, which properly simulates a mobile environment.