[FIX] voip: refactoring hiccups
During the refactoring, sanitized version of the numbers were introduced but unused in the calls.
With the addition of a faulty parameter in res.company introduced in phone_validation this would lead
to a wrong number called.

This commit ensures that the right sanitized number is used.

--------

Also on v11, incoming calls trigger a dialog. Before this commit the behaviour would be to reject the call
when closing the dialog, whether the user chose to accept or reject the call.

This commit ensures that the reject call will only happen if the session is not 'onCall'.