[FIX] l10n_mx_edi: Fix method used in unittests to compare XMLs
Currently, the method used in unittests to ensure two XMLs are equal,
evaluates as not equal an empty node and a node which contains only
spaces or newlines. For instance, the following nodes are evaluated as
not equal:

```
<node/>
```
and
```
<node>
</node>
```

This change ensures that, when a node's text contains only spaces or
newlines, they're not taken into account and are evaluated as an empty
string.