[IMP] web_studio: display correct view after closing
Before this rev. the view displayed after closing Studio was always the same
(i.e. the first view linked to the action).

Now, if you swich between views in Studio, the correct view will be opened
(exception made for the search view which can be edited with Studio but
obviously can't be opened).

Task #31963