[FIX] print: send letter
When clicking on 'Send' of the 'Send by Post' wizard, a traceback occurs
because of missing mandatory fields.

Since the read-only fields are not sent to the server, we need to use
the `force_save` attribute. Note that it is not supported in a tree
view, only in an inline tree view.

opw-805687