[FIX] hr_appraisal: stop writing on employee when creating appraisals
When an appraisal is created, the departement_id field (related on
hr.employe and stored) gets written, triggering the inverse write
on the employee.

This is unfortunate, because this write can in turn trigger a write
on analytic lines via the resource model, which will then raise the
error 'you cannot modify a validated timesheet'.

Since it does not make much sense to change the department of an
employee via the appraisal form, let's make this field readonly to
avoid triggering the inverse function of the related field.