[FIX] web_studio: render kanban record with widget
The rev. odoo/enterprise@2a996f0 replaces the function call `undelegateEvents`
by `off`. The latter expects a non-empty argument `events` which was not given,
thus causing a traceback.

Calling `off` on the widget  better suits the desired behaviour (which is
removing all events as we only want to click for edition).