[FIX] l10n_mx_edi: Fix generation of invoice's QR
According to the SAT documentation, the QR included on printed versions
of the CFDI must decodify to an URL, including specific parameters, so
the CFDI could be consulted.

Currently, the QR was a representation of the parameters, but the URL
and one parameter were missing.

This change modifies the way the QR is generated, so the full URL is now
included.

Example of an old resulting text:
?re=ACO560518KW7&tt=885.0&id=E2C75264-B462-4EBB-B353-7E6A8BC8A5FE&rr=XEXX010101000&fe=12345678

Example of a new resulting text:
https://verificacfdi.facturaelectronica.sat.gob.mx/default.aspx?re=ACO560518KW7&tt=885.0&id=E2C75264-B462-4EBB-B353-7E6A8BC8A5FE&rr=XEXX010101000&fe=12345678