[FIX] web_studio: keep action context when leaving
Since rev odoo/enterprise@3aeab29 a `do_action` is done when leaving Studio
to restore the action being edited.

The context wasn't correctly kept during this operation, which was causing
a traceback when leaving an action with a mandatory context (e.g. Helpdesk,
Tickets: `active_id` was not defined anymore).