[FIX] web_studio: always display appswitcher background
The main part of Studio frontend code is not executed if the user is
not a system user (it is put in a separate bundle since rev.
odoo/enterprise@33956b4).

This should not be the case for the code that displays the appswitcher
background otherwise the other users can't see it.