[FIX] web_mobile: have datetime picker adequate tz
In 11.0 an issue with the iOS/Android app was solved with 62ec60bd3e3.

This was backported on 10.0 with 8b340b8e but this sometimes leads to an
issue because the `get_value` is in UTC (whilst in 11.0 it was in user
local timezone).

This commit transforms the UTC date/datetime to local timezone to fix
this, and to save the gotten value do the inverse.

To parse the received value, parse_value is used because there is a
fallback on ISO 8601 date format.

forward-port-note: only for 10.0 up to saas-15 (doesn't happen in s16)

opw-803009
closes #1820