[FIX] web_enterprise: required fields in kanban are bold
When we introduced the possibility of adding sub form views in kanban
quick create, many new combinations of css classes were added at the
same time.  One problem was introduced because of the interplay between
kanban quick create css and the regular css from the enterprise version:
required fields are no longer displayed in bold.

To reproduce this issue, simply open the project application (tasks) and
quick create a task.  The main field (the title) is required, but not
displayed as such (in enterprise, the community version is fine).

With this commit, we simply remove an useless rule that caused the
issue.