[IMP] l10n_mx_edi: Solving rounding issues on cfdi 3.3
Solving the group of cases where we can not, for technical reasons,
relay into the rounding Odoo system to round the amount in the XML
itself even if it is mathematically correct on Odoo itself.

In the CFDI the rounding must be fixed into the number of decimals
mentioned in the currency by the SAT[1] (for example for MXN is
2 on sheet c_Moneda on official document [1]) but for functional and
operational reasons it is ok for us to have more decimals into Odoo
itself (3, 4 or 6 decimals even can and must be used).

The list of use cases covered are in the PR related.

PR - Closes #1829
opw806125

[1] http://www.sat.gob.mx/informacion_fiscal/factura_electronica/Documents/catCFDI.xls