[MOV] ocn_client: Odoo Cloud Notification for mobile push notifications
Push notification client for registering devices to ocn service and send
notification payload to service to notify devices (ios and android)

Before this, every single instance of odoo has their own functionality to
register and send notifications to their registered devices. For Android its
easy to handle push notification with different firebase project id for each
instances which work fine with the sing android app for different instances.
But for iOS application, we are not able to handle this situation
(i.e, different project id for each instance)

iOS required single certificate (for APN) for an application to mapped with it
and that certificate need to add in firebase project for application and that
make problem to allow every user to add their own project settings.
This is the case for on-premises databases.

So, we introduce our central platform OCN (Odoo Cloud notification) where all
instances will be registered with their user (when the user requested) and
get one token from ocn service which will be later used to send notification
payload to devices.
mail_push module is replaced by ocn client and ocn service.

After enabling push notification for instance this module performs instance
registration to ocn service and when a user on mobile application every device
of the user will be registered on ocn service mapped to their instance.

When there is data to notify the user, ocn client sends the payload with
user tokens to ocn service and ocn will take care for notifying each of
the users' devices.

Related to task #48231
Closes #2088