[FIX] web_mobile: m2m setValue with object not int
Set value of a many2many expects an object with `{id: [record_id]}` and
not just `record_id`.

In `web.field_utils` parsing many2one allows several types of data so it
works well, but the many2many is just _.identity expecting data already
alright.

But on mobile the many2one widget could be used by a many2many_tags, in
this instance the _setValue would be wrong causing an error.

This change corrects this so the _setValue also works with many2many.

opw-807577
closes #1840