[ADD] mail_enterprise:  display attachment preview in form view
This commit allows to display a preview of the chatter attachment on side of
form view if that form view in opened in large displays.
Note that it will only preview pdf and image attachments.

How it works:
* to avoid extra rpc call to fetch attachments, we are using attachment data
  available in the chatter
* we triggered attachment data from thread_field widget, and this trigger caught
  by FormRenderer and then renders attachment preview
* another issue fixed by this approach is, when new attachment added from the
  chatter, it was not displaying attachment in preview because adding new message
  in chatter does not reload form view state
* for preview we exchange the position of chatter and attachment preview node.
  With this we can manage separate scroll bars for form sheet+chatter and
  attachment preview

To use this, simply add a div with class o_attachment_preview in form view arch
to display attachment preview. Some options can be passed to change its behavior.

Options:
    types: ['image', 'pdf']
    order: 'asc' or 'desc'

e.g. if you want to display only pdf type atttachment and the latest one then use
<div class="o_attachment_preview" options="{'types':['pdf'], 'order': 'desc'}"/>

Task 40663

Co-authored-by: Martin Geubelle <mge@odoo.com>