[FIX] l10n_mx_edi: Fix URL codifycation while generating QR
Since QRs in reports are constructed from a URL
(`/report/barcode/?...`), some characters like `/` and `&` need to be
escaped. Moreover, if the text to codify into the QR is also a URL like
in this case, it's content need to be double-escaped.

This change double-escapes the QR's content, so the above mentioned
characters are codified correctly. In addition, QR's text is now passed
as a parameter (`/report/barcode?value=...`).

This has the following advantages:
- It avoids the parameters from the SAT URL to be parsed as the QR URL
ones
- If any of the RFC contains `&`, it's also escaped.
- The `//` characters from the SAT URL are handled correctly

Example of a produced SAT URL:
https://verificacfdi.facturaelectronica.sat.gob.mx/default.aspx?rr=GA%C3%910510069S1&tt=885.0&fe=MRl8zg%3D%3D&re=ACO560518KW7&id=C5B5772A-B0F5-40B8-8D9B-3B4EFCFBB613