[REF] web_studio,*: tests: calls to removeSrcAttribute
* account_reports

In the tests environment, we listen to the 'DOMNodeInserted' event
to remove the src attribute of img and iframe nodes as soon as they
are inserted into the DOM, so that they don't perform RPCs.

A community commit [REF] moves the event handler to the
addMockEnvironmnent helper, which is used by the tests in studio,
so it is no longer necessary to manually bind that handler.

Moreover, since this community commit, the event is listened on the
body instead of on the tested widget, so images appended outside
the widget are patched as well. This is why an account_reports test
had to be adapted.