[FIX] web_studio: create ACL for stages
Two ACLs are normally created when creating a new model in Studio (one for
`group_system` and one for `group_user`).

This wasn't the case for the model created for Kanban stages (see
'Enable Stages' button in Kanban editor). The `user_context` where `studio`
is supposed to be set wasn't correctly set in this case.

Another context had also been forgotten when creating an inline view.

opw-807192