[FIX] web_studio: keep context after action reload
When a view type is added to the action for example, this action is
reloaded in Studio.

The action environment (context, domain, etc.) was lost during this
reload because some preprocessing done in `doAction` was not correctly
done.

This causes an issue when adding the kanban view and trying to enable
stages.