[FIX] project_timesheet_synchro: chrome extension generation
This fix adapts the script to generate the Chrome extension
in 10.0 with the new webclient.
An optional argument is added to precise your Odoo community
repository.
Example:
./generate_extension.sh ../../odoo

This extension is a standalone app based on our framework.
The script copies some files from the framework to run the
application independently.

A major adaptation has been to keep the session active.
In 10.0, 'session_reload' no longer request the server.
'session_info' is stored in a template generated by the server.
See odoo/odoo@9c68bc748

But the extension only have static templates.
So we need to go back to the previous behavior by making a
RPC call every time the application starts.