[FIX] project_timesheet_synchro: retrieve session to stay connected
This fix is only for the desktop mode of the app.
For the 'standalone' mode See @0dd0a79a7d9

We should be connected through the active session.
But in 10.0, this behavior has changed to avoid unnecessary RPCs.
Now, we don't retrieve session_info from the server but we have to
store it into the page source.

See odoo/odoo@9c68bc7482