[FIX] l10n_mx_edi: AssertionError in Payment test task#23887
This **error** showed in runbot v11:
AssertionError: 1 != 2 : Documents not attached correctly

To **Fix** it:
The following parameters: `'default_model': 'account.payment',
'default_res_id': payment.id,'`  are added (to successfully link the
attachment build for email  to the payment) here:

wizard_mail = self.env['mail.compose.message'].with_context({
'default_template_id': default_template.id}).create({})

The `l10n_mx_edi_retrieve_attachments()` method is added,
because it provides the same function as we were doing with a
whole new search in payment model.

The assert check if two XML are attached: one build when the
payment is validated and the second one build when the email
message is created.