[FIX] l10n_mx_edi: QR generation
Since QRs in reports are constructed from a URL
(`/report/barcode/?...`), some characters like `/` and `&` need to be
escaped. Moreover, if the text to codify into the QR is also a URL like
in this case, it's content need to be double-escaped.

This change double-escapes the QR's content, so the above mentioned
characters are codified correctly. In addition, QR's text is now passed
as a parameter (`/report/barcode?value=...`).

On the other hand, this change also causes the stamp (or "sello")
characters to be included as they are, withoud beeing escaped. This is
because, even though the stamp contains special characters like "=", the
SAT expects exactly eight unescaped characters on the URL.

This has the following advantages:
- If any of the RFC contains `&`, it's escaped, which adds support for
RFCs containing such character.
- The characters "=", "+" and "/", which may be part of the stamp, are
preserved as it should.
- The `//` characters from the SAT URL are not reduced to just one "/"
and are handled correctly
- It prevents the parameters from the SAT URL to be parsed as the QR URL
ones

Example of a URL produced before this change:
https://verificacfdi.facturaelectronica.sat.gob.mx/default.aspx?rr=BC&041101JV0&tt=885.0&re=ACO560518KW7&id=9F5E9A02-9CC7-450D-A50A-2B72AC34D43A&fe=4%j2%2BA%3D%3D

Example of a URL produced after:
https://verificacfdi.facturaelectronica.sat.gob.mx/default.aspx?rr=BC%26041101JV0&tt=885.0&re=ACO560518KW7&id=9F5E9A02-9CC7-450D-A50A-2B72AC34D43A&fe=4/j2+A==

Closes #1851
opw-815445