[REF] helpdesk: remove assigned subtype and keep only classic tracking
Since https://github.com/odoo/odoo/commit/e9734f4975904b6196423b7f5de7dd211809d6f6 assignation is now a notification
sent to the inbox or by email to the newly-assigned responsible. It is
therefore considered as not necessary to link a subtype to the assignation
itself. Indeed having it logged in the discussion history as a classic
tracking in addition to the notification is considered as sufficient.

This commit is related to task ID 60790. Closes #1896 .