[FIX] web_enterprise: mention manager not cropped
When the mention manager is in a scrollable element, what is over the
top/left of that element will be hidden.

This happen really easily on enterprise desktop version since the
chatter is on the left and the composer is at the top of a scrollable
element.

This fix in the case of the sidebar chatter position the dropdown below
the composer avoiding the issue.

opw-814562
closes #1898