[IMP] project_timesheet_forecast_sale: plan project when timesheet only
The code of the project overview is split to be used when only
timesheet is installed (no forecast). The table part only
display column for timesheeted time per month.
When forecast is installed and activated, other columns
comes to complete this table.

Most of the removed code goes in community (sale_timesheet
module).