[FIX] barcodes: Do not trigger useless onchange server side.
The purpose is to improve barcode performance. Since there
is no onchange server side for picking, inventory and lot
wizard we can skip the server call (needed function are directly
call with an xmlrpc).

This commit is only able to block the onchange on the parent object
since options(notifyChange, setDirty,..) are not propagated on x2many
records (modify this behavior could have side effect for example on
handle widget with resequence.)

Howerver on an inventory with 535 inventory lines (without lot)
we win 2000 ms for the inventory object onchange on a field total_qty
that is only used for the inventory report.