[FIX] web_studio: give a name to nameless groups/pages
* Go to some form view in studio, do some modifications
* Open the XML editor and inside any xpath manually introduce at least
    two unnamed groups
* Manually put at least one element inside each of the unnamed groups, save
* Place another element within the view with studio, which will trigger
    the normalization

Result? All of the content inside any unnamed group after the first
unnamed group is put back into the first unnamed group

Why? During normalization, we assumed that if a node has the same
attributes, tag, etc than another node then they're the same node, this
is not the case!

Solution? If we're parsing a group and or a page (containers) and it
doesn't have a name attribute, we assign it one :^)

OPW-814773