[FIX] stock_barcode: key press qty
This commit complete the lot wizard and picking view in
order to be able to enter quantity done after scan.
for picking:
-the code in order to select and remember candidate
was already present in the view and model, only the key
was missing in widget.

for lot:
-setQuantityWithKeypress was already set thus there was a
 clear purpose to do it, however the code server side with
 the wizard was not adapted.

This commit add all the missing element in order to enable
the js for key press.

commit linked d54491f74512958d6e7bc76eb065ca8999bb673e