[FIX] web_studio: create x2m and m2o fields
... without specifying a related field (for one2many) or a relation
(for many2one and many2many).

Enter studio for a given model, open the form view, drag and drop a
new x2m or m2o field: a dialog opens. The user must select a
related field or a relation.

Before this rev., if he clicked on 'Confirm' without selecting one,
it crashed. This rev. displays a warning and keeps the dialog
opened.

Note that the case was already handled for related fields, but we
uniformized the way the warning is displayed.

Fixes #23241