[FIX] quality_mrp: header no overflowing in tablet
On the work order tablet view it is easy depending on content and
resolution / orientation of tablet to get a header bar with overflowing
content.

The content can still be scrolled to, but this is then white text on
white background which make this worse.

With this commit, when the header bar is overflowing items on the right
will be placed on a second line.

opw-1816335
closes #1918