[FIX] project_timesheet_synchro: generate mobile bundle
Before this commit, this script didn't work because of some
refactoring and due to mobile tools changes (Cordova, SDK, Gradle,..).

This fix can be considered as an improvement but it is important to
generate the mobile app properly after a fix.
Anyway, this is a tool script only used at our own.

We also had to add manually unversioned app icons to the Cordova
project. These icons are now automatically generated from a single icon.