[FIX] project_timesheet_synchro: avoid absolute path
Due to this recent commit @a6fa7a786d, the mobile apps
didn't work after generating these.
project_timesheet_synchro has to work in four different
environments: Android, iOS, Chrome extension, Iframe.

(1)
We cannot use absolute paths because they're different on all
platforms, and there are no plugins or configurations to unify
absolute paths.
Example with the window.location.pathname value of the Android app:
"/android_asset/www/index.html"
As we can see, the root directory ('/') refers to 'android_asset'
and doesn't correspond to the root project ('www').
So, depending of the env., we are never certain to know the location
of the master directory.

(2)
Because of (1), we decide to use relative path but
we need to handle a special case with the iframe.
Because of the iframe src (/project_timesheet_synchro/timesheet_app),
the current path location is 'project_timesheet_synchro' instead of
the root path.
This is why we added a new variable named 'rootPath' who is
the relative path to the root directory.