[FIX] project_timesheet_forecast_sale: only assigned timesheet on overview
A timesheet can be created without employee_id set, since it is
not required, as a timesheet is an analytic line. Thus, when trying
to display employee name (of False) in overview, it crashes.
We then restrict the timesheet to only the ones with an employee
set to avoid error, in forecast/timesheet table.