[FIX] account_batch_deposit: missing trigger for amount computation
In a batch deposit,
when changing the amount of a payments
by going in the payments (Sales > Payments),
cancelling the payment, changing its amount
and re-confirm it,
the total amount of the deposit was not updated.

This could be seen in the batch deposit list amount,
not in the deposit itself, where the total amount
displayed is the actual sum of all deposit lines.

opw-1819744