[FIX] quality: Don't show message if failure if type is passfail
The user is able to create a failure message for quality
point with test type 'pass or fail'. However this message
is only display when a user fail a measure for a quality
point with a test type 'take a measure' (basic behavior
introduced by commit 8a0f904)

This commit adds a domain in order to hide the 'message if
failure' page if the type is passfail without a workorder step
set on the quality point.

This commit will introduce an error if a user update quality_mrp
without update quality. We decided to do it because this problematic
should be less important than having a confusing field for the user.
Also this behavior will be removed in master and instead it will display
the failure message in a pass/fail check.

opw-1819910