[FIX] account_sepa: show bank accounts of the commercial partner
When a SEPA payment was done from the invoice form view, the contact
partner would be used, and his bank accounts were used to register the
payment.
The domain on the view enforced that no other accounts could be used.
The bank accounts may only be defined for the commercial partner of the user,
so we change the domain when selecting the SEPA payment method to display both
the accounts of the contact and his commercial partner.

Co-authored by: Nicolas Martinelli <nim@odoo.com>
opw-1814650