[FIX] sale_subscription: avoid multiple subs at upsell/renewal
Before this commit, order lines in renewal or upsell orders created
from a subscription would not be linked to the subscription in question,
which would cause the creation of a second subscription containing the
extra products at confirmation instead of placing all products in the
subscription one wished to renew/upsell.

While the proper way to go would be to store the id of the subscription
being worked on somewhere on the SO to avoid polluting the source field,
this is not something we can do in a stable release, hence the 'hacky'
feeling one might feel along their spine when reading this. However, I
like to comfort myself by saying that the 'origin' field is in fact
functionaly relevant in that case since the subscription is indeed the
origin of the order.

Also, tests \o/