[FIX] l10n_mx_edi: Use up to 6 decimals in the currency rate in the CFDI
Currently, in the CFDI, the attribute rate (`TipoCambio`) takes its
precision from the currency being used. However, according to the SAT,
the precision does not depend on the currency, and must have up to 6
decimal digits

This change modifies the precision for that attribute and set it to 6
decimals.

In addition, this changes the expected rate in the unittests, in order
to match the produced number of decimals