[ADD] l10n_mx_edi: Add field to store maximum of allowed currency decimals
According to the SAT, amounts in the CFDI must have a specific maximum
of decimal digits, which depends on the currency of that amount. For
this purpose, the SAT provides a catalog, where a list of currencies are
provided along with their number of allowed decimal digits (named
`c_moneda`).

This change adds a new field to the currency model to store the maximum
number of decimal digits, and adds the data provided by the SAT's
catalog.

The following currencies are not included, because there is a mismatch
between the currencies supported in the SAT's catalog and the ones
supported by Odoo:
- SDG: was put out of circulation in 2007, SDD is now used.
- SRD: not supported yet in Odoo, still supporting SRG, even though the
currency change took place in 2004
- TMT: not supported yet in Odoo, still supporting TMM, even though the
currency change took place in 2009
- ZMW: not supported yet in Odoo, still supporting ZMK, even though the
currency change took place in 2012
- ZWL: not supported yet in Odoo, still supporting ZWD