[IMP] l10n_mx_reports: Added date filter in COA report
Added filter to define the date to set in the filename and in the XML to
generate.

The XML have the attribute YEAR and MONTH, that indicate the period in
which was generated the document. Currently always take the date when is
generated the XML, but is not the correct, because must corresponding
with the other reports.

Now is possible assign the period, and this will be assigned in the XML