[FIX] l10n_mx_edi_cfdi_33: Support tax 0%
When a invoice only have tax 0%, the node ´TotalImpuestosTrasladados="0.00"´

Then when I validate `taxes['total_transferred'] is not None` in this
case always is 0, and the attribute is not added.

Now, better validate by the key that have the taxes description, that
only is fill when the invoice have taxes.

** Fixes the error:

 El valor del campo TotalImpuestosTrasladados no es igual a la suma de
 los importes registrados en el elemento hijo Traslado