[FIX] web_enterprise: color of hovered badges
By default, the color of bootstrap's badges on hover is white,
which isn't adequate on white backgrounds, obviously. In the
enterprise webclient, we already override their base color to be
the 'Odoo green', so in this rev., we simply define their color
on hover to be a darker version of that green.