[FIX] web_studio: add and remove groups on fields in views
The groups set in views were not correctly displayed in the sidebar after the
rev. odoo/enterprise@4d7024b which is basically reverted in this commit.
Trying to set the attribute `studio_groups` after the `_apply_group` execution
doesn't make any sense as the `groups` attribute is deleted.

The initial issue was the RNG validation which doesn't accept the
`studio_groups` attribute. However, this preprocess is only useful in the Studio
`fields_view_get`, not when creating/writing on the view.

Fixes #23243