[ADD] l10n_mx_edi_landing: Completed l10n_mx_edi_custom module
This module extends l10n_mx_edi_custom, to allow register the customs
number in a landing, and assign automatically the custom number in the invoice
lines when is a customer invoice with a imported product

To get this value, is required all the flow. Must be generated the
purchase, receive the products, create the landing related, after create
the sale order, send the product and generate the invoice.

For more detail in the module README.