[IMP] sale_subscription: Add renewal options on subscription template
Purpose
=======

Renewal should be an option in subscription templates. That way we can hide a lot of stuff by default (end date, renewal buttons).

Specification
=============

Radio button Period: Unlimitted vs Time bounding
    if Limited: duration field (same than recurrency)
    if Limited: show and compute End date in subscription
Hide renewal buttons if the option is not activated in the template

This commit is related to task ID 34831