[FIX] inter_company_rules: description of generated SO line
When creating a SO in the other company on purchase order
confirmation, we want the sale item description to be the
same of its origin PO line.
The inverse is already done, this commit uniformize the
behavior, as we think if the description contains something
specific, this information should be transmit to the destination
company.

opw-1824498