[FIX] web_studio: add a related field
One could not create a related field pointing to a selection or
relation (many2one or x2many) field with Studio.

This is a backport of rev. odoo/enterprise@05027c8 and partial one of rev.
odoo/enterprise@effb095

Note that the latter one could not be applied because it depends on the rev.
odoo/odoo@3e63f01 which cannot be backported in a stable version ; a warning
message is thus displayed for one2many related field.

opw-1826288