[FIX] timesheet_grid: better precision for reminder
If activated, the emails reminder for employees and
managers are sent for a certain period. However, the
period is strange, since it starts with a date and
end with a datetime (cron execution datetime). This leads
to some "not so friendly" message, like "you only
timesheeted 3 hours out of the expected 25.52".

The period depends on the cron execution time, which is
not intuivitve and correct.

This commit makes calculations  use date instead of
datetime to compute the interval to check, and includes
entirely the date_stop (even if the day is not completely
over) in the period.

opw-1826026