[FIX] web_enterprise: style required m2m [10-s15]
Many2many tags widget in list view were not shown as required when they
were.

In the form view, the bottom border is present and in community the blue
background is shown in form or list view for for many2many tags widget.

note: for 10.0 up to saas-15 (in saas-16 the style has been changed and
also work for many2many tags widget)

opw-1814298
closes #1999