[FIX] web_studio: edit form view without record
When there is no record to edit the form view, we make a default one
out of default values. This first implemention of the future basic
model cannot handle a list of commands as default m2m values.

This was causing an issue while editing the Product form view without
default values for the taxes.

Instead of trying to handle the list of commands, which doesn't bring
any value in Studio, we just ignore the default values.

opw-1826338