[FIX] web_gantt: only apply first group_by to first group by
In the gantt view, if for example we group by 3 fields, and have 2
records in 2 first level grouping, split_groups will be called eg.:

 1 time for the first grouping level (=> alright)
 2 times for each grouping after the first grouping level (=> alright)
 3 times for each grouping after second grouping level (=> bad, should be 1)
 3 times for each grouping after third grouping level (=> bad, should be 1)
 1 time for each grouping after fourth grouping level (=> alright)

So here the function could be called 45 times instead of 9.

This happened because if we have 2 records:
 (write_uid1, create_uid1, name1),
 (write_uid2, create_uid2, name2)

that we group by write_uid, create_uid, name.

The gantt would have first level grouping:

 {write_uid1, write_uid2}

and for second level grouping:

 {write_uid1, write_uid2, create_uid1, create_uid2}

 instead of {create_uid1, create_uid2}

and third level grouping:

 {write_uid1, write_uid2, name1, name2}

 instead of {name1, name2}

If we have a use case with more grouping and group by, this can explode
and make the code really resources expensive. For example if we grouped
by 8 fields with 16 records, it would be possible to have hundred
millions groupings instead of a hundred that in the end will be the only
ones to be shown (since only them contain a record to show).

This is not needed in 11.0 since that code is dead since refactoring.

opw-1816590
closes #2005